backend/requirements.txt
sanic==22.12.0
sanic-ext==22.12.0
sanic-testing==22.12.0
pymongo==4.10.1
python-dotenv==1.0.1
google-generativeai
pytest
